God knows how much decoupling and refactoring just happened.  Trying to make the renderers abstracted and also making the ownership more clearly defined...there was WAY too many classes going outside the bounds of their implied contracts.  Next up: changing from a simulation to able to be real-time or not.
