+++ index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Ironhack Mars Rover</title>
+  </head>
+  <body>
+   <h3>Ironhack Mars Rover</h3>
+   
+   <canvas id="canvas" width="288" height="512"></canvas>
+   
+   <script src="mars-rover.js"></script>
+  </body>
+</html>
